[project @ 8a5ef5ec7f56f69a0cec187676755940ef2cee6c] You should never _ever_ do something like this:
       [process A]  ---[piped write until EOF]---> [process B]
       [process A]  <---[piped read until EOF]--- [process B]

Basically, what happens is something like this: A feeds
data to B; B processes the data, and feeds it to the pipe
until the pipe buffer is full; at this point, A writes to
its pipe until its buffer is full too, and if it still has
more data to write, that's it, it never gets to the read
part to flush B's buffer --- both processes are blocked on
write.

This patch fixes such a situation in the compose menu.

(From: Liviu Daia <daia@stoilow.imar.ro>)

Original author: Thomas Roessler <roessler@does-not-exist.org>
Date: 1998-07-21 07:18:48+00:00

git-svn-id: http://oss.sygroup.ch/svn/mutt/trunk@184 2bb92213-1a62-4c20-bdc6-606bde718ed5