[project @ 595194d021a24d5e79e2c69d4e04f89db14e7f82] There is a long-standing problem in Mutt, related to coloring the various levels of quoting: if the attributions are missing, like this:         >>> blah blah from A 	>> blah blah from B 	> blah blah from C
then the "quoted" color object is associated with ">>>", "quoted1"
with ">>" and "quoted2" with ">" --- which is not what most people
expect. The reason is Mutt doesn't count the quote characters (since
there is no way to distinguish between a single quote ">> " and a
">" followed by a "> "), and it allocates colors as it finds new
types of quote prefixes. The attached patch fixes this problem still
without counting the quote characters.

Original author: Thomas Roessler <roessler@does-not-exist.org>
Date: 1998-09-23 06:03:55+00:00

git-svn-id: http://oss.sygroup.ch/svn/mutt/trunk@225 2bb92213-1a62-4c20-bdc6-606bde718ed5