[project @ 865abb7d9ffaa2014386eaa6d29aef8075a55788] Properly use tables in manual where appropriate.
Now, lots of information which really deserves a table is just put into
<screen/> sections with manual line breaks and things like that.

This is ugly for several reasons including that the information stored
there should be accessable straight from the table of contents.

For the TOC, this patch switches to <book/> for the manual and uses a
deeper split level so things actually look nicer.

The only problem are tables which may be not so well supported when
dumping the HTML to .txt but the DocBook->txt converter with table
support is work-in-progress...

Original author: Rocco Rutte <pdmef@gmx.net>
Date: 2006-05-18 18:02:34+00:00

git-svn-id: http://oss.sygroup.ch/svn/mutt/trunk@4013 2bb92213-1a62-4c20-bdc6-606bde718ed5