[project @ acd71f2f255506824d3a00511d159607dee63034] Fix RFC 3676 (format=flowed text) handling.
The old code would consider a line containing "> " to be flowed, but
since this is a quoted and space-stuffed line containing no additional
text, by my reading of RFC 3676 it should be fixed.

Clean up the handling of format=flowed text.  Fix the test to determine
whether a line is fixed--if a line ends in a space only because the last
character is a space from space-stuffing, consider the line to be a
fixed line.  This makes the test for ((buf_len - buf_off) <= 0) later no
longer necessary.

Also simplify the code by removing checks for curline being non-null; it
is allocated at the start of the function and never reallocated to size
zero, so it should never be a null pointer.

Original author: Michael Vrable <mvrable@cs.ucsd.edu>
Date: 2007-08-15 20:09:42-07:00

git-svn-id: http://oss.sygroup.ch/svn/mutt/trunk@4494 2bb92213-1a62-4c20-bdc6-606bde718ed5