[project @ 78aab353e217082241e048bd7042f90e0036e2b9] Only consider empty files non-existent for buffy's purpose.
According to SUS, the st_size member of struct stat is undefined
for directories. In case a system reports 0, the newly_created
flag will be set for a directory and the magic will be cleared.

If the folder type is cleared, mutt won't check the folder for
new mail and thus will ignore MH and Maildir type folders on
such a fs completely (e.g. the HAMMER fs on DragonFly BSD).

Original author: YONETANI Tomokazu <qhwt+mutt@les.ath.cx>
Date: 2008-08-29 10:23:28+02:00

git-svn-id: http://oss.sygroup.ch/svn/mutt/trunk@4779 2bb92213-1a62-4c20-bdc6-606bde718ed5